requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
click==8.1.7
Django==5.0
django-debug-toolbar==4.2.0
django-filter==23.5
django-rq==2.10.1
djangorestframework==3.14.0
Markdown==3.5.2
psycopg2==2.9.9
pytz==2024.1
redis==5.0.1
rq==1.15.1
sqlparse==0.4.4
typing_extensions==4.8.0
requests==2.31.0
